refactor(mobile start points): decouple drawSession from angular
